refine Clang script

+++ scripts/buildclangfromcross.sh
@@ -14,6 +14,8 @@ echo "-> Setting up directories"
 
 # build directory
 mkdir -p $BUILD_DIR/LLVM-Clang
+mkdir -p $PREFIX/licenses
+mkdir -p $PREFIX/cleanup
 
 # build LLVM-Clang
 echo "-> Building LLVM/Clang"
@@ -22,24 +24,39 @@ cd $BUILD_DIR/LLVM-Clang
 
 # copy licenses
 echo "-> Copying LLVM/Clang licenses"
-mkdir -p $PREFIX/licenses
 cd $PREFIX/licenses
-mkdir -p LLVM
-cp $SRC_DIR/LLVM/CREDITS.TXT LLVM/CREDITS.TXT
-cp $SRC_DIR/LLVM/LICENSE.TXT LLVM/LICENSE.TXT
-mkdir -p clang
-cp $SRC_DIR/LLVM/tools/clang/LICENSE.TXT clang/LICENSE.TXT
+if [ -f licenses.marker ]
+then
+  echo "--> Licenses already installed."
+else
+  echo "--> Installing LLVM/Clang licenses."
+  mkdir -p LLVM
+  cp $SRC_DIR/LLVM/CREDITS.TXT LLVM/CREDITS.TXT
+  cp $SRC_DIR/LLVM/LICENSE.TXT LLVM/LICENSE.TXT
+  mkdir -p clang
+  cp $SRC_DIR/LLVM/tools/clang/LICENSE.TXT clang/LICENSE.TXT
+  echo "--> Done!"
+fi
+touch licenses.marker
 
 # copy environment setup
 cp $TOP_DIR/envsetup/clang32env.cmd $PREFIX/
 
 # cleanup
 echo "-> Cleanup"
-cd $PREFIX
-echo "--> Removing libtool files"
-find . -name \*.la -exec rm -f {} \;
-echo "--> Stripping executables"
-find . -name \*.exe -exec $HOST-strip {} \;
+cd $PREFIX/cleanup
+if [ -f cleanup.marker ]
+then
+  echo "Already cleaned up."
+else
+  cd $PREFIX
+  echo "--> Removing libtool files"
+  find . -name \*.la -exec rm -f {} \;
+  echo "--> Stripping executables"
+  find . -name \*.exe -exec $HOST-strip {} \;
+  echo "--> Done!"
+fi
+touch cleanup.marker
 
 # zipping
 echo "-> Packaging Clang addon package"
